refactor: stamp csp nonces through the router instead of rewriting html
The document-header middleware read the whole response body so it could
regex `<script` tags into carrying a nonce. That implementation is a
fossil of `scripts/serve.mjs`, a static file server with no router
handle, where string rewriting was the only option available; the Nitro
migration carried it into `start.ts` unchanged.

TanStack Router has supplied the hook the whole time. Setting
`router.options.ssr.nonce` makes `Scripts`, `HeadContent` and the SSR
serializer stamp the nonce onto every tag they emit, and
`renderRouterToStream` passes the same value to React for its bootstrap
and suspense-resolution scripts. `HeadContent` also renders the
`<meta property="csp-nonce">` tag the client reads it back from, so a
client-side navigation keeps injecting scripts the policy accepts.

Nothing upstream mints a nonce — Router only ever reads that option — so
the value is still ours to generate. `@server/csp` does it once per
request, keyed on the request object so concurrent renders cannot share
one, and both the router and the header middleware read the same value
whichever asks first.

The middleware is now headers-only and the body streams straight
through, which is what lets a route with a slow loader render
progressively rather than waiting on a fully buffered document. The
`content-length`/`content-encoding`/`transfer-encoding` deletions go
with it: they only existed to describe a body that `response.text()`
had already decoded.
