+++ apps/web/src/router.tsx
@@ -3,6 +3,7 @@ import { readSentryDsn } from "@app/sentryDsn";
 import LoadingPlaceholder from "@base/LoadingPlaceholder";
 import RouteError from "@base/RouteError";
 import * as Sentry from "@sentry/tanstackstart-react";
+import { getRequestNonce } from "@server/csp";
 import { MutationCache, QueryCache, QueryClient } from "@tanstack/react-query";
 import { createRouter } from "@tanstack/react-router";
 import { getCommonOptions } from "@virtool/sentry/browser";
@@ -62,6 +63,16 @@ export function getRouter() {
 		scrollRestoration: true,
 		scrollToTopSelectors: [`#${CONTENT_SCROLL_ID}`],
 	});
+
+	// Router and React stamp this nonce onto every script they emit, including
+	// the inline ones a CSP's `script-src 'self'` cannot cover: the dehydration
+	// payload and React's suspense-resolution frames. Router only ever reads
+	// this option, so supplying the value is ours to do. `getRouter` runs once
+	// per request, so this is the request's own nonce — the same one the
+	// document-header middleware names in `script-src`.
+	if (import.meta.env.SSR) {
+		router.options.ssr = { nonce: getRequestNonce() };
+	}
 
 	// `import.meta.env.SSR` is a compile-time constant, so the whole block is
 	// dead-code-eliminated from the server bundle. `browserProfilingIntegration`
