+++ docs/auth.md
@@ -49,6 +49,7 @@ not per-handler. The exempt endpoints live in
 export const authenticationExceptions: ReadonlyArray<{ url: string }> = [
   createFirstUserFn,
   getPasswordPolicyFn,
+  getRootFn,
   loginFn,
   logoutFn,
   resetPasswordFn,
@@ -58,16 +59,30 @@ export const authenticationExceptions: ReadonlyArray<{ url: string }> = [
 and are wired up in `apps/web/src/start.ts`:
 
 ```ts
-const authenticationMiddleware = createAuthenticationMiddleware(
-  authenticationExceptions,
-);
+const authenticationMiddleware = createAuthenticationMiddleware();
 
 export const startInstance = createStart(() => ({
   defaultSsr: false,
-  requestMiddleware: [csrfMiddleware, cspNonce],
-  functionMiddleware: [authenticationMiddleware],
+  serializationAdapters: [serverErrorSerializationAdapter],
+  requestMiddleware: [
+    sentryGlobalRequestMiddleware,
+    metricsMiddleware,
+    csrfMiddleware,
+    documentHeadersMiddleware,
+  ],
+  functionMiddleware: [
+    sentryGlobalFunctionMiddleware,
+    errorLoggingMiddleware,
+    authenticationMiddleware,
+  ],
 }));
 ```
+
+`createAuthenticationMiddleware` takes no list in production — it
+reaches `./exceptions` through a `createServerOnlyFn` dynamic import on
+first call, so `start.ts`, which is part of the browser program, never
+drags those functions' modules into the eager client bundle. The
+parameter exists only so tests can supply their own list.
 
 Every server function is gated by default. Public endpoints opt out by
 being listed in the `exceptions` array — passed as **server-function
